Refactor: Editor components into sub components
